backend/codeleap_backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
gunicorn>=22.0
whitenoise>=6.6